refactor(brutus): extract execution runners to runner.go
Move runFromStdin and runSingleTarget functions.
Part of main.go refactoring (8/9).

Related: C1/C2 (split 1340-line main.go)
