+++ cmd/brutus/runner.go
@@ -1,0 +1,273 @@
+// Copyright 2026 Praetorian Security, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"bufio"
+	"encoding/json"
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/praetorian-inc/brutus/internal/plugins/snmp"
+	"github.com/praetorian-inc/brutus/pkg/brutus"
+)
+
+// runFromStdin reads fingerprintx JSON from stdin and tests each target
+func runFromStdin(base *baseConfigOptions, jsonOut bool) ([]brutus.Result, bool) {
+	var allResults []brutus.Result
+	hasSuccess := false
+
+	scanner := bufio.NewScanner(os.Stdin)
+	for scanner.Scan() {
+		line := strings.TrimSpace(scanner.Text())
+		if line == "" {
+			continue
+		}
+
+		// Parse fingerprintx JSON
+		var fpx FingerprintxResult
+		if err := json.Unmarshal([]byte(line), &fpx); err != nil {
+			if base.useColor {
+				fmt.Fprintf(os.Stderr, "%s%s Warning:%s failed to parse JSON: %v%s\n", ColorYellow, SymbolWarning, ColorReset, err, ColorReset)
+			} else {
+				fmt.Fprintf(os.Stderr, "Warning: failed to parse JSON: %v\n", err)
+			}
+			continue
+		}
+
+		// Determine protocol: use override if specified, otherwise map from fingerprintx
+		var protocol string
+		if base.protocolOverride != "" {
+			protocol = base.protocolOverride
+		} else {
+			protocol = mapServiceToProtocol(fpx.Protocol)
+			if protocol == "" {
+				// Unsupported service, skip
+				continue
+			}
+		}
+
+		// Determine TLS mode for this specific target
+		targetTLSMode := base.tlsMode
+		if targetTLSMode == "disable" {
+			if metadata, ok := fpx.Metadata["tls"]; ok {
+				// If TLS is detected in metadata, auto-upgrade to skip-verify
+				if tlsEnabled, ok := metadata.(bool); ok && tlsEnabled {
+					targetTLSMode = "skip-verify"
+					if base.verbose {
+						fmt.Fprintf(os.Stderr, "[verbose] TLS detected in fingerprintx metadata, auto-upgrading to skip-verify mode\n")
+					}
+				}
+			}
+		}
+
+		// Build target string
+		target := fmt.Sprintf("%s:%d", fpx.IP, fpx.Port)
+
+		// AI mode: For HTTP services, detect auth type and route appropriately
+		if base.aiMode && (protocol == "http" || protocol == "https") {
+			useHTTPS := protocol == "https"
+			authType, banner := detectHTTPAuthTypeWithBanner(target, useHTTPS, base.timeout, base.verbose)
+			if authType == "basic" {
+				// HTTP Basic Auth detected - use LLM to research credentials
+				if base.verbose {
+					fmt.Fprintf(os.Stderr, "[verbose] AI mode: %s uses Basic Auth, using LLM analysis\n", target)
+				}
+
+				// Research credentials using LLM
+				if base.llmConfig != nil && base.llmConfig.Enabled {
+					creds := researchCredentialsWithLLM(target, banner, base.llmConfig, base.verbose)
+					if len(creds) > 0 {
+						// Add researched credentials to base config for this target
+						base.aiResearchedCreds = creds
+						if base.verbose {
+							fmt.Fprintf(os.Stderr, "[verbose] LLM researched %d credential pairs for %s\n", len(creds), target)
+						}
+					}
+				}
+				// Keep http/https protocol
+			} else {
+				// No Basic Auth - likely form-based, use browser protocol
+				if base.verbose {
+					fmt.Fprintf(os.Stderr, "[verbose] AI mode: %s appears form-based, using browser automation\n", target)
+				}
+				protocol = "browser"
+			}
+		}
+
+		// Temporarily set TLS mode for this target, then restore
+		originalTLSMode := base.tlsMode
+		base.tlsMode = targetTLSMode
+
+		// Run against this target
+		results, success := runSingleTarget(target, protocol, base)
+
+		// Restore original TLS mode to prevent mutation across targets
+		base.tlsMode = originalTLSMode
+		allResults = append(allResults, results...)
+		if success {
+			hasSuccess = true
+		}
+
+		// Output valid credentials immediately (streaming for large-scale scans)
+		if !jsonOut {
+			outputValidOnly(results, base.useColor)
+		}
+	}
+
+	if err := scanner.Err(); err != nil {
+		if base.useColor {
+			fmt.Fprintf(os.Stderr, "%s%s Error:%s reading stdin: %v%s\n", ColorRed, SymbolError, ColorReset, err, ColorReset)
+		} else {
+			fmt.Fprintf(os.Stderr, "Error reading stdin: %v\n", err)
+		}
+	}
+
+	return allResults, hasSuccess
+}
+
+// runSingleTarget runs brutus against a single target
+func runSingleTarget(target, protocol string, base *baseConfigOptions) ([]brutus.Result, bool) {
+	config := &brutus.Config{
+		Target:        target,
+		Protocol:      protocol,
+		Usernames:     base.usernames,
+		Passwords:     base.passwords,
+		Keys:          base.keys,
+		UseDefaults:   true,
+		NoBadkeys:     !base.useBadkeys,
+		Threads:       base.threads,
+		Timeout:       base.timeout,
+		StopOnSuccess: base.stopOnSuccess,
+		LLMConfig:     base.llmConfig,
+		TLSMode:       base.tlsMode,
+		RateLimit:     base.rateLimit,
+		Jitter:        base.jitter,
+		MaxAttempts:   base.maxAttempts,
+		SprayMode:     base.sprayMode,
+	}
+
+	// Handle SNMP-specific tier selection (tiers are CLI-only, not in library defaults)
+	if protocol == "snmp" && len(config.Passwords) == 0 {
+		if !snmp.ValidateTier(base.snmpTier) {
+			if base.useColor {
+				fmt.Fprintf(os.Stderr, "%s%s Error:%s invalid --snmp-tier: %s (use: default, extended, full)%s\n",
+					ColorRed, SymbolError, ColorReset, base.snmpTier, ColorReset)
+			} else {
+				fmt.Fprintf(os.Stderr, "Error: invalid --snmp-tier: %s (use: default, extended, full)\n", base.snmpTier)
+			}
+			return nil, false
+		}
+		config.Passwords = snmp.GetCommunityStrings(snmp.Tier(base.snmpTier))
+	}
+
+	// Handle HTTP with AI-researched credentials
+	if (protocol == "http" || protocol == "https") && len(base.aiResearchedCreds) > 0 {
+		// Clear default usernames/passwords - use only AI-discovered paired credentials
+		config.Usernames = nil
+		config.Passwords = nil
+
+		// Add AI-researched credentials
+		config.Credentials = append(config.Credentials, base.aiResearchedCreds...)
+
+		// Add admin:admin as basic fallback
+		config.Credentials = append(config.Credentials, brutus.Credential{
+			Username: "admin",
+			Password: "admin",
+		})
+
+		if base.verbose {
+			fmt.Fprintf(os.Stderr, "[verbose] Using %d AI-researched credentials for HTTP (+ admin:admin fallback)\n", len(base.aiResearchedCreds))
+		}
+
+		// Disable LLM in config to prevent duplicate analysis in brutus.go
+		config.LLMConfig = nil
+
+		// Clear researched creds so they don't leak to next target
+		base.aiResearchedCreds = nil
+	}
+
+	// Handle browser-specific configuration
+	// Browser protocol requires AI mode - it's the AI-powered form authentication feature
+	if protocol == "browser" {
+		config.Threads = base.browserTabs
+		config.Timeout = base.browserTimeout
+
+		// Clear default usernames/passwords - browser uses AI-discovered paired credentials only
+		config.Usernames = nil
+		config.Passwords = nil
+
+		// Use Claude Vision + Perplexity for credential research
+		// Also get the configured browser plugin to use during brute force
+		aiCreds, browserPlugin := researchBrowserCredentials(target, base)
+		if len(aiCreds) > 0 {
+			config.Credentials = append(config.Credentials, aiCreds...)
+			if base.verbose {
+				fmt.Fprintf(os.Stderr, "[verbose] AI researched %d credentials for browser\n", len(aiCreds))
+			}
+		}
+		// Pass the configured plugin to Brute (so it has VisionAnalyzer, AIVerify, etc.)
+		if browserPlugin != nil {
+			config.Plugin = browserPlugin
+		}
+	}
+
+	// Verbose: print config summary before starting
+	if base.verbose {
+		fmt.Fprintf(os.Stderr, "[verbose] Target: %s (protocol: %s)\n", target, protocol)
+		fmt.Fprintf(os.Stderr, "[verbose] Paired credentials: %d, Usernames: %d, Passwords: %d, Keys: %d\n",
+			len(config.Credentials), len(config.Usernames), len(config.Passwords), len(config.Keys))
+		// Calculate total attempts
+		totalAttempts := len(config.Credentials)
+		if len(config.Passwords) > 0 {
+			totalAttempts += len(config.Usernames) * len(config.Passwords)
+		}
+		if len(config.Keys) > 0 {
+			totalAttempts += len(config.Usernames) * len(config.Keys)
+		}
+		fmt.Fprintf(os.Stderr, "[verbose] Total attempts: %d, Threads: %d, Timeout: %s\n",
+			totalAttempts, config.Threads, config.Timeout)
+		fmt.Fprintf(os.Stderr, "[verbose] Starting brute force...\n")
+	}
+
+	// Run brute force
+	results, err := brutus.Brute(config)
+	if err != nil {
+		if base.useColor {
+			fmt.Fprintf(os.Stderr, "%s%s Error:%s testing %s: %v%s\n", ColorRed, SymbolError, ColorReset, target, err, ColorReset)
+		} else {
+			fmt.Fprintf(os.Stderr, "Error testing %s: %v\n", target, err)
+		}
+		return nil, false
+	}
+
+	// Check for success
+	hasSuccess := false
+	successCount := 0
+	for i := range results {
+		if results[i].Success {
+			hasSuccess = true
+			successCount++
+		}
+	}
+
+	// Verbose: print completion summary
+	if base.verbose {
+		fmt.Fprintf(os.Stderr, "[verbose] Completed: %d results, %d successful\n", len(results), successCount)
+	}
+
+	return results, hasSuccess
+}
